+++ src/codemagic/tools/app_store_connect.py
@@ -492,25 +492,41 @@ class AppStoreConnect(
         "register-device",
         BundleIdArgument.PLATFORM,
         DeviceArgument.DEVICE_NAME,
-        DeviceArgument.DEVICE_UDID,
+        DeviceArgument.DEVICE_UDIDS,
+        DeviceArgument.IGNORE_REGISTRATION_ERRORS,
     )
     def register_device(
         self,
         platform: BundleIdPlatform,
         device_name: str,
-        device_udid: str,
+        device_udids: Types.DeviceUdidsArgument,
+        ignore_registration_errors: bool = False,
         should_print: bool = True,
-    ) -> Device:
+    ) -> List[Device]:
         """
-        Register a new device for app development
+        Register new Devices for app development
         """
-        return self._create_resource(
-            self.api_client.devices,
-            should_print,
-            udid=device_udid,
-            name=device_name,
-            platform=platform,
-        )
+        registered_devices = []
+        for i, device_udid in enumerate(Types.DeviceUdidsArgument.resolve_value(device_udids)):
+            if should_print:
+                self.echo("") if i != 0 else None
+
+            try:
+                device = self._create_resource(
+                    self.api_client.devices,
+                    should_print,
+                    udid=device_udid,
+                    name=device_name,
+                    platform=platform,
+                )
+            except AppStoreConnectError as error:
+                if not ignore_registration_errors:
+                    raise error from error
+                self.logger.error(Colors.YELLOW(f"Failed to register a device: {error.args[0]}"))
+            else:
+                registered_devices.append(device)
+
+        return registered_devices
 
     @cli.action(
         "create-bundle-id",